Implemented the Class Choice dialog which allows the player to choose between three ways to define a class.
Added a generic infobox dialog which displays some text and a set of buttons (vertical only for now), this is used for the Class Choice dialogs and the dialogs which shows various questions the player must choose from. The questions are currently hardcoded.
Added more gui states to handle the extra class dialogs (Generate, Pick or Create).