Use debug dlls when debugging in vs2010 (try 2)
Using the Debug build in vs2010 is not working because the debug dlls are
not loaded when debugging. The reason they are not loaded is that
CMAKE_BUILD_TYPE is not defined when doing multiple builds. This in turns
causes OGRE_PLUGIN_DEBUG_SUFFIX not to be set. This patch makes sure that
OGRE_PLUGIN_DEBUG_SUFFIX is always set but only used when debugging.

It also defines DEBUG to make it easier turn things on and off when
debugging.

There are still other bugs that have broken Debug mode in vs2010 but those
will be addressed in other patches.