added hack to fix an alignment issue in MSVC 2010
The default allocator in Visual Studio 2010 does not respect the
alignment requirements of classes it is allocating memory for. This
results in a potential crash when using OEngine::Physics::PhysicActor
that has been allocated on the heap if it is built against a version of
Bullet that has SSE intrinsics enabled.