Revert "Use a child scene node for the accumulation root"
This reverts commit d6f923f2743afd4bd781e5a31bc5f897f3f18785.

We don't need it for any of the NIFs we're currently handling. As long as
there's no NIF files that would break it, we should require a stationary root
if an animation wants to accumulate. If we must, a better idea may be to inject
an extra bone into the skeleton instance and make that the accumulation root.