Introduce -DBUILD_WITH_DPKG to toggle dpkg based install
The current system automagically chooses between the dpkg-based install method
(for debian-derived distributions) and a "traditional" install and sets
install paths based on whether cmake can find a 'dpkg' executable.

This is not ideal, since dpkg is occasionally installed on linux distributions
unrelated to debian for purposes other than package management. In particular,
Arch and Gentoo carry it in their repositories.