Center, switch and change text for confirmation buttons
The confirmation buttons for quitting or starting a new game were
driving me crazy by not being ok left and cancel right, as e.g. vanilla
is.

Then I noticed the text isn't the same as in vanilla either. It should
be yes/no, not ok/cancel.

To make it even more authentic I also centered the buttons. Had to
remove autosized buttons for this. I think it looks fine still.

Comments?