Right aligned version/revision
I've always felt that having the version/revision text at the bottom
center in the main menu was a bit out of place. A more common place for
this kind of thing is in one of the corners. I chose bottom right.

Aditionally I right aligned it and changed the v and r in version and
revision to capital letters.

Comments?