Fix "unknown info ID" exceptions when a dialogue response is from the Info Refusal group
Could be observed in TG_OverduePayments quest when talking to Trasteve about Dwemer Artifacts.

Info Refusal responses are not specific to any particular topic, so they should not be added to the journal. Trying to do so anyway will cause "unknown id" exceptions because MWDialogue::Entry expects the infoId to be from the Dialogue for the supplied topic.