Update in response to comments.
It turns out I was erroneous about problem (1) in my original commit
message. I thought that in an edge case the OpenMWGUI could both handle
the same event on the frame the GUI was closed. My test for
this was broken. So after brainstorming many possible unsatisfactory
solutions to this event handling, it's luckily not necessary!