Add a full search to findInteriorPositionInWorldSpace.
Part of OMW Bug #1533
Implement a search for one of the 'nearest' exterior cells.  In this
case, 'nearest' means the fewest number of cells away via door markers.
This causes the world map position to update immediately after teleporting,
unless the new cell has no connecting path to an exterior.  Intervention
spells and Jail travel will be much closer to vanialla Morrowind,
except for in Mournhold.