mannequins: final fix (hopefully !)
finally the 2nd kind of mannequins was the one with a script attached
containing a skipanim call. The problem in openmw is that killDeadActors is
called when mSkipAnim has been set to false, after processing the normal
update for an actor.
So for now I have put a work-around, mSkipAnim is set to false after
killDeadActors and killDeadActors can test mSkipAnim state.