ultimate fix for walk on water effect
it just removes the test for the level at which waterCollision works.
It's even better than vanilla this way because in vanilla you suddenly
appear on top of water when using this, here you are pushed to the surface
and you see the movement. With this walk on water works everywhere and no
risk to get out of the cell by the top as before !