Always set a voice stream as initially pending
This keeps the voice stream starting code to a common location.

This also avoids a race condition where a loudness object and decoder can be
set up for async loading, and get loaded before it's checked for being ready,
causing the loudness data to be ready and the decoder be used without being
rewound.