Update convert_bump_mapped_mods.rst
Ok, so another update then. Sorry for another one so soon after the first PR, but I think (or hope) that this will be the last edit for some time. It looks good to me now. ´

Mainly, I changed changed some "=" headers to "-", and while at it, I fixed some typing errors and just edited some text to read better.