Add Visual Studio 2017 support
The new VS2017 configuration uses all the same settings as the
VS2015 one, but uses a different generator.

CMake by default looks for Boost libraries based on the compiler
being used, but Boost only has binaries for VS2015, so a hack is
used to make sure it looks for the correct ones.

I don't know what TOOLSET and XP_TOOLSET are for but I left them
just in case.