Combine stacks of similar items when vendor sells items (bug #2842)
When merchants have an item in an owned container, and an identical item
on their person, those stacks do not appear combined in the trading
screen.  This is because the items in the container may not have their
ownership set on a per-item basis, but the items on the vendor's person
do.  When trying to combine stacks, the code would see that the owners
didn't match, and would not join them.

Since the code to build the container model for the trader already
checks that the items are owned by the vendor, it is safe for the
model to assume that if it is trading, then item ownership can be
ignored for the sake of stacking.