fix CI msvc build outside git bash
It was only by accident that git bash worked for CI basing on the
"real_pwd()" expression. Replace the dubious replacement pattern with
"cygpath", which is present on git bash, Cygwin, and msys2 alike. In
particular git bash uses msys2 internally.

I was able to confirm a working build under msys2 with a wrapper for
MSVC invocation via the Ninja generator.