De-pseudonymise myself on the authors list
I'm the author of commit 01283f5 and #151. At that time I was just about to start a new job, and unsure how open-source contributions would be viewed by my boss. Therefore I made the pull request (but stupidly enough not the commit) under a pseudonymous account, which has since been deleted.
In the end my worries were baseless, and I would like to be listed with my full name on the authors list.