Split volume into volume factor and SFX volume
Volume factor is passed as an argument when engine logic requires to play
sound. SFX volume is loaded from ESM file. These values are not available
at the same time and can change independently. With asynchronous sound loading
SFX volume becomes available when access to the volume factor argument is lost
therefore need to keep it as a separate field.