simplify node allocations
Currently, we use a somewhat strange approach to replace an osg::Group node we just created with a SceneUtil::Skeleton node. With this PR we create the correct type of node in the first place. As a side effect of these changes, NightDayLabel and HerbalismLabel will now be applied properly when a Skeleton node is used.